[BACK-777] More caching fixes
- Updated deleteCollectionStory mutation to fetch the entire story object
back now that BACK-802 is fixed. My hope was that it would render refetching
getCollectionStories query unnecessary but it appears still to be needed when
you add or delete stories.

- Now refetching queries when new authors and collections are added so that
they appear on their respective list pages.

- Disabled the 'Status' field and set it to 'Draft' by default on the 'Add Collection'
page so that a newly created, story-less collection cannot be published by accident.