[BACK-747] Implement image uploads
- Added react-dropzone package to leverage easy uploads.

- Updated Modal component to be more generic (the original version
was only going to be used for displaying a loading message)

- On the Author page, removed the "Upload new image" button and instead
made the image/placeholder itself clickable. A click on the image brings up
a modal with a (yet to be styled) droppable area for image uploads. A click
on this droppable area will bring up the OS file upload menu.

The rather proof-of-concept implementation shows available file info for the
file about to be uploaded.