[BACK-832] Improve UX around story image uploads
- Updated ImageUpload component to have an action button at the bottom, clicking
on the image itself also works as before.

- Updated 'add a story to collection flow' - now we first save a story, then
upload an image to S3 and if that worked out, we save it in the story record.

- If the Parser doesn't return a 'topImageUrl', we use the first image in the image
array if there is one - this fixes the YouTube thumbnails issue (among perhaps outhers?).

- If the Pocket Image Cache cannot process the image for us to sidestep CORS, adding
a story to a collection is no longer blocked.

- Misc UI improvements: larger story thumbnails in desktop view, edit/delete buttons are
stacked vertically to free up a bit of horizontal space, Add/Edit story now has a 'Cancel'
button - when adding a story, this resets the form, and when editing, it toggles it back
into hidden state.