[BACK-104] Show list of curated items
This required a surprising amount of prep work, since it's the first query to be used
with the new Curated Corpus API. There is some minor refactoring - the /config folder
is now in /src and not hidden within collections.

- Added a route and a new CuratedItemsPage that pulls in a list of curated items - no
filters or custom ordering used yet, just pagination.

- Added a component to show each curated item with a link to a (future) page. Added tests
for the new component.