[BACK-1169] Show a display of items for a selected new tab
- Added a "Schedule" page for items scheduled on new tab. By default, this page
shows scheduled entries for today (in the user's local time zone) and tomorrow.

There are buttons above and below the displayed items to load more scheduled items,
two days' worth at a time.

- Realising that the cards displayed on the Schedule page are identical to the ones
used on the Live Corpus page, I refactored the ApprovedItemListCard component to remove
the action buttons and place it in an ApprovedItemCardWrapper component. That way, I could
reuse the card within a wrapper specific to the Schedule page: ScheduledItemCardWrapper.

- Added tests for new components and updated tests for components that have been updated
(specifically, the ApprovedItemListCard now displays flags for syndicated, collection
and short-lived items).