feat(schedule): [BACK-1170] Wire the "Remove" button on Schedule page
- Added a new mutation to be able to delete a scheduled item from New Tab.

- Updated types, set up a new modal, form with a single checkbox that asks
to confirm the removal. Given the potential gravity of this decision I thought
just the "Confirm" button is not enough, so added a checkbox.

- Added tests for the new form.

- Stopped using NewTabGroupedList to display scheduled entries as it meant
I would have had to feed the "onRemoveSave" function to the scheduled card
through a couple of components that had no need for this at all.

- Now that NewTabGroupedList is no longer needed on the Schedule page, I stripped the
switching between card components from it but am still hanging on to it in case it
comes in handy elsewhere in the app.