feat(forms): Add progress bars
We've always had provisions for displaying when the form has been
submitted but the result of the mutation is yet to return, however
progress bars didn't get copied over from Collections initially.

More recently, these were added into the Schedule and CuratedItem
forms, and this PR adds them to the remaining forms.

Also, removing linting for YAML files as the Super Linter fails
files previously found to be just fine.