fix(editing): Articles not having topics assigned to them [BACK-1335] (#737)
Narrowed down the initial bug report to multi-word topic names:
instead of saving `HEALTH_FITNESS` on the backend, for example,
the frontend was sending through `HEALTH & FITNESS` to the API,
capitalising the display name instead of using the enum code from
the shared data.

- Updated the `topic` select in the `ApprovedItemForm` component,
and made similar changes to the `language` and `curationStatus`
selects. Updated tests to check both option values and display
values for all three fields.