feat(ui): [BACK-1331] Make items with a recomendation status visually distinct (#745)
* feat(ui): [BACK-1331] Make items with a recomendation status visually distinct

Removed the recommendation/corpus list item in the item card and added an overlay
with the word "Recommendation" on images for recommended items instead.

Moved "Topic" and "Language" properties to the bottom of the card, so that they're always
there. Moved "Syndicated", "Collection", "Time sensitive" props up. Removed dividers and
secondary colour so that there is less visual clutter.

* Remove transparency and shorten copy on the overlay

* Fix unit test following UI update + remove dot in copy