feat(mui5): [BACK-1729] Transition Landing page and shared components to MUI 5
- Updated the landing page.

- Worked through all the components in the `_shared` folder and updated them
to use MUI 5.

- `Header` was the most labour-intensive to update; there was a lot of rearranging
code in this component and some styles we definitely need (e.g. different header
background for dev environments) have to be applied at the very last moment with the
`sx` property which does not result in the cleanest of components. The upside is
that we don't touch this component often.