feat(mui5): [BACK-1757] Transition Corpus Item page to MUI 5
This change encompasses updates to the main forms and form action
components for "Edit", "Schedule" and "Reject" actions in addition
to UI updates to the Corpus Item page itself.

As such, the most work was to tame the date pickers used on the
Schedule Item form - using MUI's x-date-pickers package proved to be
the easier option of the two (the mui/lab package is being removed
with this PR as that was the second option, and we don't use Lab
components for anything else).

Tests interacting with the date pickers have been updated; an unfinished
integration test has been updated but not yet finished as that will
come in a separate test sweep PR.