feat(mui5): Transition all components from MUI 4 to MUI 5 (#986)
* feat(mui5): [BACK-1694] Initial package upgrade + first page with form (#970)

* feat(mui5): [BACK-1694] Initial package upgrade + first page with form

* Remove a test that is no longer relevant

* feat(mui-5): [BACK-1718] MUI 5 Upgrade: Collection Authors pages and theme upgrade (#973)

* feat(mui-5): [BACK-1718] MUI 5 Upgrade: Collection Authors pages and theme upgrade

- Updated all Collection Author-related components and pages to use MUI 5 packages.

- Updated the app theme to use MUI 5 themes, too. Some of our customised components
(MUI Button, Switch) are not yet restored in this new theme version - that code
is temporarily commented out.

- Where JSS styles were used previously, I've switched to using inline styles
with the `sx` property which is the recommended MUI 5 way for one-off styling.

* feat(mui): [BACK-1720] MUI 5 Upgrade: Transition Collection Search, home and landing pages (#975)

- Minimal updates required here, so upgraded one of the shared components
alongside the pages: `HandleApiResponse`.

* feat(mui5): [BACK-1719] Transition Collection Partners pages and components to MUI 5 (#974)

- Updated all Collection Partners-related code to use MUI 5 libraries

- Abstracted out some of the more complex (e.g. not one or two line) custom styles
into styled component wrappers that are now reused on both Partners and Authors list
cards.

* feat(mui5): [BACK-1721] Transition Collection pages and components to MUI 5 (#976)

* feat(mui): [BACK-1721] Transition Collection pages and components to MUI 5

- Updated all components within `src/collections` to MUI 5. This involved
a few styling/markup tweaks, updates to props in Autocomplete components
used for labels, and deletion of separate styles files with JSS styles.

* Fix focus background on clicking on cards

* Remove code that is no longer needed in tabs

* feat(mui5): [BACK-1747] Migrate Curated Corpus Rejected page to MUI 5 (#978)

- Updated the components used on this page & verified everything still works!

* feat(mui5): [BACK-1757] Transition Corpus Item page to MUI 5 (#981)

This change encompasses updates to the main forms and form action
components for "Edit", "Schedule" and "Reject" actions in addition
to UI updates to the Corpus Item page itself.

As such, the most work was to tame the date pickers used on the
Schedule Item form - using MUI's x-date-pickers package proved to be
the easier option of the two (the mui/lab package is being removed
with this PR as that was the second option, and we don't use Lab
components for anything else).

Tests interacting with the date pickers have been updated; an unfinished
integration test has been updated but not yet finished as that will
come in a separate test sweep PR.

* feat(mui5): [BACK-1729] Transition Landing page and shared components to MUI 5 (#977)

- Updated the landing page.

- Worked through all the components in the `_shared` folder and updated them
to use MUI 5.

- `Header` was the most labour-intensive to update; there was a lot of rearranging
code in this component and some styles we definitely need (e.g. different header
background for dev environments) have to be applied at the very last moment with the
`sx` property which does not result in the cleanest of components. The upside is
that we don't touch this component often.

* feat(mui5): [BACK-1755] Transition Schedule page to MUI 5 (#983)

- Updated the SchedulePage component and almost all other
components used on this page to use MUI 5 (SplitButton
has had updates made in #982, so I have not updated it here).

- Made a few visual tweaks - DatePicker fields are now slim
(with smaller padding around the data) so that their vertical
width now matches input fields on forms and buttons.

- Updated page container width to use a default MUI value for
width ("xl") instead of a fixed pixel width.

- Fixed tests for the Header element.

* feat(mui5): [BACK-1754] Transition Corpus page to MUI 5 (#984)

- Updated the Corpus page and all related components to use MUI 5.

- Removed old style files and updated broken tests.

- Updated a few components that were left behind in previous PRs,
e.g. CuratedCorpusLandingPage, RemoveItemFromScheduledSurfaceForm.

* feat(mui5): [BACK-1756] Transition Prospecting page to MUI 5 (#982)

* transition Prospecting page and all associated components to MUI 5

* tweak, clean up and address pr comments

* remove MUI 4 packages and fix tests

* feat(mui5): [BACK-1769] Transition Labels page to MUI 5 (#987)

- Updated LabelListPage and LabelListCard components to work with MUI 5.

* feat(mui5): [BACK-1770] Apply small UI tweaks (#989)

- Fixed small UI tweaks listed in the ticket - these are mostly
issues around padding and spacing between elements.

- Fixed size of prospect titles on mobiles.

Co-authored-by: Herraj Luhano <hluhano@mozilla.com>